Added multithreading for chunk loading and improved terrain algorithm

+++ SOLIS/Assets/Scripts/Planet/PlanetSettings.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [CreateAssetMenu()]
 public class PlanetSettings : ScriptableObject
 {
-    public long planetSeed;
+    public int planetSeed;
     public float planetFeatureSize;
     public Color terrainColor;
     public Color waterColor;
